Move to_graph to graph module
